backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic-settings==2.5.2
supabase==2.28.3
httpx==0.27.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.146.0
